Re-org project so everything is under `cmd` or `pkg`.
Also moved each command into its own package to make it easier to declare flags
without worrying about global variables.
