+++ cmd/config/config.go
@@ -1,11 +1,12 @@
-package cmd
+package config
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/wearefair/gurl/config"
+	"github.com/wearefair/gurl/pkg/config"
+	"github.com/wearefair/gurl/pkg/log"
 )
 
-var configCmd = &cobra.Command{
+var ConfigCmd = &cobra.Command{
 	Use:   "config",
 	Short: "Configure gurl",
 	Long: `Set configurations for gurl, which will be saved at ~/.gurl/config.
@@ -15,16 +16,12 @@ var configCmd = &cobra.Command{
 	Run: configure,
 }
 
-func init() {
-	RootCmd.AddCommand(configCmd)
-}
-
 func configure(cmd *cobra.Command, args []string) {
 	config.Read()
 	config.Prompt()
 	conf := config.Instance()
 	err := config.Save(conf)
 	if err != nil {
-		logger.Fatal(err.Error())
+		log.Logger().Fatal(err.Error())
 	}
 }
