+++ pkg/util/uri_test.go
