--- util/pointerify.go
@@ -1,5 +1,0 @@
-package util
-
-func PointerifyBool(b bool) *bool {
-	return &b
-}
